+++ openfisca_france_indirect_taxation/variables/consommation/consommation_menages.py
@@ -48,26 +48,6 @@ class depenses_ticpe(YearlyVariable):
     def formula(menage, period, parameters):
         taux_plein_tva = parameters(period.start).imposition_indirecte.tva.taux_de_tva.taux_normal
         return menage('depenses_ht_ticpe', period) * (1 + taux_plein_tva)
-        # This is equivalent to call directly poste_07_2_2_1_1
-
-
-class depenses_essence_recalculees(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Dépenses en essence recalculées à partir du prix HT'
-
-    def formula(menage, period, parameters):
-        taux_plein_tva = parameters(period.start).imposition_indirecte.tva.taux_de_tva.taux_normal
-        depenses_sp_e10_ht = menage('depenses_sp_e10_ht', period)
-        depenses_sp_95_ht = menage('depenses_sp_95_ht', period)
-        depenses_sp_98_ht = menage('depenses_sp_98_ht', period)
-        depenses_super_plombe_ht = menage('depenses_super_plombe_ht', period)
-        return (
-            depenses_sp_e10_ht
-            + depenses_sp_95_ht
-            + depenses_sp_98_ht
-            + depenses_super_plombe_ht
-            ) * (1 + taux_plein_tva)
 
 
 class depenses_tot(YearlyVariable):
@@ -119,15 +99,3 @@ class duree_moyenne_trajet_aller_retour_teg(YearlyVariable):
     value_type = float
     entity = Menage
     label = "Durée moyenne de l'aller-retour pour le teg"
-
-
-# class quantite_diesel(YearlyVariable):
-#     value_type = float
-#     entity = Menage
-#     label = "Quantité de diesel consommée (en hecto-litres)"
-
-
-class quantite_supercarburants(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Quantité de supercarburants (super 95, super98 et superE10) consommée (en hecto-litres)'
