+++ openfisca_france_indirect_taxation/variables/consommation/quantites_energies_logement.py
@@ -21,19 +21,6 @@ class quantites_combustibles_liquides(YearlyVariable):
         return quantite_combustibles_liquides
 
 
-class quantites_diesel(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Quantités de diesel consommées par les ménages'
-
-    def formula(menage, period, parameters):
-        depenses_diesel = menage('depenses_diesel', period)
-        diesel_ttc = parameters(period.start).prix_carburants.diesel_ttc
-        quantites_diesel = depenses_diesel / diesel_ttc * 100
-
-        return quantites_diesel
-
-
 # Not used
 class quantites_electricite_3kva(YearlyVariable):
     value_type = float
@@ -140,33 +127,6 @@ class quantites_electricite_18kva(YearlyVariable):
         quantite_elect = depenses_sans_part_fixe / prix_unitaire_elect
 
         return quantite_elect
-
-
-class quantites_essence(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = "Quantités d'essence consommées par les ménages"
-
-    def formula_1990(menage, period):
-        quantites_sp95 = menage('quantites_sp95', period)
-        quantites_sp98 = menage('quantites_sp98', period)
-        quantites_super_plombe = menage('quantites_super_plombe', period)
-        quantites_essence = (quantites_sp95 + quantites_sp98 + quantites_super_plombe)
-        return quantites_essence
-
-    def formula_2007(menage, period):
-
-        quantites_sp95 = menage('quantites_sp95', period)
-        quantites_sp98 = menage('quantites_sp98', period)
-        quantites_essence = (quantites_sp95 + quantites_sp98)
-        return quantites_essence
-
-    def formula_2009(menage, period):
-        quantites_sp95 = menage('quantites_sp95', period)
-        quantites_sp98 = menage('quantites_sp98', period)
-        quantites_sp_e10 = menage('quantites_sp_e10', period)
-        quantites_essence = (quantites_sp95 + quantites_sp98 + quantites_sp_e10)
-        return quantites_essence
 
 
 class quantites_gaz_contrat_base(YearlyVariable):
@@ -305,64 +265,3 @@ class quantites_gaz_liquefie(YearlyVariable):
         quantites_gaz_liquefie = depenses_gaz_liquefie * total_gpl_phebus / total_gpl_bdf
 
         return quantites_gaz_liquefie * 1000  # en KWh
-
-
-class quantites_sp_e10(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Quantités consommées de sans plomb E10 par les ménages'
-
-    def formula(menage, period, parameters):
-        depenses_essence = menage('depenses_essence', period)
-        part_sp_e10 = parameters(period.start).imposition_indirecte.part_type_supercarburants.sp_e10
-        depenses_sp_e10 = depenses_essence * part_sp_e10
-        super_95_e10_ttc = parameters(period.start).prix_carburants.super_95_e10_ttc
-        quantite_sp_e10 = depenses_sp_e10 / super_95_e10_ttc * 100
-
-        return quantite_sp_e10
-
-
-class quantites_sp95(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Quantités consommées de sans plomb 95 par les ménages'
-
-    def formula(menage, period, parameters):
-        depenses_essence = menage('depenses_essence', period)
-        part_sp95 = parameters(period.start).imposition_indirecte.part_type_supercarburants.sp_95
-        depenses_sp95 = depenses_essence * part_sp95
-        super_95_ttc = parameters(period.start).prix_carburants.super_95_ttc
-        quantite_sp95 = depenses_sp95 / super_95_ttc * 100
-
-        return quantite_sp95
-
-
-class quantites_sp98(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Quantités consommées de sans plomb 98 par les ménages'
-
-    def formula(menage, period, parameters):
-        depenses_essence = menage('depenses_essence', period)
-        part_sp98 = parameters(period.start).imposition_indirecte.part_type_supercarburants.sp_98
-        depenses_sp98 = depenses_essence * part_sp98
-        super_98_ttc = parameters(period.start).prix_carburants.super_98_ttc
-        quantites_sp98 = depenses_sp98 / super_98_ttc * 100
-
-        return quantites_sp98
-
-
-class quantites_super_plombe(YearlyVariable):
-    value_type = float
-    entity = Menage
-    label = 'Quantités consommées de super plombé par les ménages'
-
-    def formula(menage, period, parameters):
-        depenses_essence = menage('depenses_essence', period)
-        part_super_plombe = \
-            parameters(period.start).imposition_indirecte.part_type_supercarburants.super_plombe
-        depenses_super_plombe = depenses_essence * part_super_plombe
-        super_plombe_ttc = parameters(period.start).prix_carburants.super_plombe_ttc
-        quantite_super_plombe = depenses_super_plombe / super_plombe_ttc * 100
-
-        return quantite_super_plombe
